requirements.txt
aiolifx
aiohttp
colour
click
PyYAML
